Create local IslandManager
* Created component for common code between the managers
* Renamed EvvoIslandActor -> RemoteEvvoIsland for consistency
