+++ src/main/scala/com/evvo/remoting/Remoting.scala
@@ -6,7 +6,7 @@ import akka.actor.ActorSystem
 import com.evvo.agent.{CreatorFunc, MutatorFunc}
 import com.evvo.agent.defaults.DeleteWorstHalfByRandomObjective
 import com.evvo.island.population.{Maximize, Objective}
-import com.evvo.island.{EvvoIslandActor, IslandManager, StopAfter}
+import com.evvo.island.{RemoteEvvoIsland, RemoteIslandManager, StopAfter}
 import com.evvo.{CreatorFunctionType, MutatorFunctionType}
 import com.typesafe.config.ConfigFactory
 
@@ -25,7 +25,7 @@ object Remoting {
 
   def main(args: Array[String]): Unit = {
 
-    val builder = EvvoIslandActor.builder[Int]()
+    val builder = RemoteEvvoIsland.builder[Int]()
       .addObjective(Objective[Int](x => x, "identity", Maximize))
       .addCreator(CreatorFunc(creator, "Set(0)"))
       .addMutator(MutatorFunc(mutator, "Add1"))
@@ -38,7 +38,7 @@ object Remoting {
 
     implicit val actorSystem: ActorSystem = ActorSystem("EvvoNode",config)
 
-    val islandManager = IslandManager.from(10, builder)
+    val islandManager = new RemoteIslandManager(10, builder)
 
     islandManager.runBlocking(StopAfter(1.second))
